Move files around for better structure.
